Instruction Rig
Here we setup the foundation for how the Butler would carry out the
instructions to fulfil intent. This revision sees the butler acting on
every instruction as they are given from the command. It needs not
buffer; it needs not remember. Every intent the user has, gets passed on
as an instruction which the Butler acts on.

My original thoughts here are detailed in the Dispatcher. I wanted
something which incorporated event sourcing. I think event sourcing is a
better argument here because we would like to know how a project
progressed through time. Without event sourcing, we would only have the
snapshot of the project in time. This would mean, however, that I need
to setup the event sourcing rig... ain't nobody got time for that; at
least not right now. But if I had all the time in the world to design
this before first launch, definitely event sourcing. Decoupling reasons.

The sequence I settled for, in this revision, looks something like this

+-----------+      +------------+      +-------------+
| CLI Input |----->| Dispatcher |----->| Instruction |
+--+--------+      +------------+      +-------------+
   | Intent |             |                    |
   +--------+             |    +-----------+   |
                          +--->| Exception |<--+
                               +-----------+

The Dispatcher serves as a router for the Butler. Since my mind wouldn't
shut up, it's claiming that this could be some sort of broker service at
scale; but we aren't at scale yet.

The Instruction is the quintessential command. This is what the Butler
must do. At the moment. This action is mapped to the intent which comes
from the CLI Input.

Exceptions are how the Butler flips the table and goes "ain't nobidy got
time for this". But of course, in a polite way. A Butler must be polite
afterall.