Base initialization of a Butler
Here we handle creating the directory the butler would work out of.
Nothing special; just a hidden directory.

The good old "run the executable as a test" showed that exceptions
weren't handled properly in the code. Fixing this, I realized Crystal
may be as flexible, if not more flexible than Ruby if one knows how to
bend the rules properly.

I cannot remember if this was possible in Ruby — because I never tried
— but Crystal allows you scope tests within the module as opposed to
having your tests litter the global scope. I find this fascinating. This
means if modules are well written within a package, it would be easier
to rip them out in the future into their own services.