Add the ability to add tasks to the Butler (#3)
Here we add some tasks to the Butler. Man's gotta know what to keep
track of. This is a simple "create a card on Trello"; so nothing fancy.
But there are some design decisions I am slowly realizing needs to be
made.

- Storage: currently, this is in a flat file as JSON, but it could be
something more elaborate as time goes on and the entity space grows.
- TDD: TBH, IDGAF. I'd rather build this out first... then create a test
harness later. That's just how my brain works. — although, I had to
write tests for the Store at some point cos mans didn't understand what
Crystal was doing behind the scenes. Turns out all the test you see for
the Store module was to prefix the expression in `Store#unique?` with !
- Sugar: `task add` is a backward way of speaking. I realize this is how
programmers speak, but Butlers are more polished. I would have to
introduce some sugar to this. What I currently don't know is how this
plays with the help text.
- Exceptions: I was just an ass here. I don't know why I tried to
maintain "errors" and "exceptions"

Writing to the datastore is an atomic task. This is could become a
problem if the number of tasks become ginormous cos I'm handling this in
one process. But when we get to that bridge, we'd cross it. i cannot
come and go and kill myself. Here's the note that I am aware of it.

At this point, I can add tasks. But there's nothing I can do with this
ability because Butler neither knows how to show me these tasks, nor how
to interact with them.

Worklog:

* Put tests in modules; don't litter the global scope
* Better routing + Task instruction
* Fixed logger write bug
* Add the Task entity
* Add a minimal store
* Add "CreateTask"
* Fixed tests setup and teardown