List Tasks (#6)
For this we are looking to view the tasks we added using create task.
I am going for the very minimal with this implementation. Thouts in my
head before this was to have some elaborate list with boxes properly
laid out. But I realized too quickly that that adds to time, but does
not really add value to me. So that's a flurry for another day.

* Add the route to the Dispatcher
* Fix spec name - unrelated
* Add time to log
* Make tasks human-readable
* List tasks